pyproject.toml
[project]
name = "lisa-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "streamlit-aggrid>=1.1.6",
    "supabase>=2.16.0",
]
